# Optimize generic types (2)