--- a/column_buffer_amd64.go
+++ b/column_buffer_amd64.go
@@ -1,0 +1,6 @@
+//go:build !purego
+
+package parquet
+
+//go:noescape
+func writeValuesInt32(values []int32, rows array, size, offset uintptr)