--- a/column_buffer_amd64.go
+++ b/column_buffer_amd64.go
@@ -3,4 +3,7 @@
 package parquet
 
 //go:noescape
+func writeValuesBits(values []byte, rows array, size, offset uintptr)
+
+//go:noescape
 func writeValuesInt32(values []int32, rows array, size, offset uintptr)