--- a/column_buffer_amd64.go
+++ b/column_buffer_amd64.go
@@ -13,6 +13,9 @@ func writeValues32bits(values unsafe.Pointer, rows array, size, offset uintptr)
 //go:noescpae
 func writeValues64bits(values unsafe.Pointer, rows array, size, offset uintptr)
 
+//go:noescape
+func writeValues128bits(values unsafe.Pointer, rows array, size, offset uintptr)
+
 func writeValuesBool(values []byte, rows array, size, offset uintptr) {
 	writeValuesBitpack(*(*unsafe.Pointer)(unsafe.Pointer(&values)), rows, size, offset)
 }
@@ -31,6 +34,10 @@ func writeValuesUint32(values []uint32, rows array, size, offset uintptr) {
 
 func writeValuesUint64(values []uint64, rows array, size, offset uintptr) {
 	writeValues64bits(*(*unsafe.Pointer)(unsafe.Pointer(&values)), rows, size, offset)
+}
+
+func writeValuesUint128(values []byte, rows array, size, offset uintptr) {
+	writeValues128bits(*(*unsafe.Pointer)(unsafe.Pointer(&values)), rows, size, offset)
 }
 
 func writeValuesFloat32(values []float32, rows array, size, offset uintptr) {