--- a/column_buffer_amd64.s
+++ b/column_buffer_amd64.s
@@ -139,40 +139,6 @@ loop8x4:
     CMPQ SI, CX
     JE done
 
-    /*
-    CMPQ CX, $4
-    JB loop1x4
-
-    MOVQ CX, DI
-    SHRQ $2, DI
-    SHLQ $2, DI
-
-loop4x4:
-    MOVL (BX), R8
-    ADDQ DX, BX
-
-    MOVL (BX), R9
-    ADDQ DX, BX
-
-    MOVL (BX), R10
-    ADDQ DX, BX
-
-    MOVL (BX), R11
-    ADDQ DX, BX
-
-    MOVL R8, (AX)(SI*4)
-    MOVL R9, 4(AX)(SI*4)
-    MOVL R10, 8(AX)(SI*4)
-    MOVL R11, 12(AX)(SI*4)
-
-    ADDQ $4, SI
-    CMPQ SI, DI
-    JNE loop4x4
-
-    CMPQ SI, CX
-    JE done
-    */
-
 loop1x4:
     MOVL (BX), R8
     MOVL R8, (AX)(SI*4)