--- a/column_buffer_purego.go
+++ b/column_buffer_purego.go
@@ -1,0 +1,9 @@
+//go:build !amd64 || purego
+
+package parquet
+
+func writeValuesInt32(values []int32, rows array, size, offset uintptr) {
+	for i := range values {
+		values[i] = *(*int32)(rows.index(i, size, offset))
+	}
+}