This PR adds vector optimizations for routines that write to column buffers.

Comparisons to using the pure Go versions of the algorithms:
```
name                                          old time/op  new time/op  delta
GenericBuffer/benchmarkRowType/go1.18         1.49µs ± 0%  0.75µs ± 0%   -50.06%  (p=0.000 n=10+9)
GenericBuffer/booleanColumn/go1.18             589ns ± 0%   316ns ± 0%   -46.39%  (p=0.000 n=9+9)
GenericBuffer/int32Column/go1.18               576ns ± 0%   235ns ± 0%   -59.16%  (p=0.000 n=9+10)
GenericBuffer/int64Column/go1.18               554ns ± 0%   305ns ± 0%   -44.96%  (p=0.000 n=9+9)
GenericBuffer/floatColumn/go1.18               616ns ± 0%   235ns ± 0%   -61.77%  (p=0.000 n=10+10)
GenericBuffer/doubleColumn/go1.18              616ns ± 0%   305ns ± 0%   -50.45%  (p=0.000 n=9+8)
GenericBuffer/byteArrayColumn/go1.18          9.91µs ± 1%  9.94µs ± 2%      ~     (p=0.781 n=10+10)
GenericBuffer/fixedLenByteArrayColumn/go1.18  5.45µs ± 0%  5.50µs ± 0%    +1.00%  (p=0.000 n=10+9)
GenericBuffer/stringColumn/go1.18             6.21µs ± 0%  6.26µs ± 0%    +0.75%  (p=0.000 n=9+8)
GenericBuffer/indexedStringColumn/go1.18      40.2µs ± 0%  40.1µs ± 1%      ~     (p=0.370 n=8+9)
GenericBuffer/uuidColumn/go1.18                888ns ± 0%   391ns ± 0%   -56.03%  (p=0.000 n=10+10)
GenericBuffer/mapColumn/go1.18                 693µs ± 0%   705µs ± 1%    +1.71%  (p=0.000 n=9+10)
GenericBuffer/decimalColumn/go1.18             577ns ± 0%   306ns ± 0%   -46.87%  (p=0.000 n=10+10)
GenericBuffer/contact/go1.18                  13.0µs ± 1%  12.9µs ± 0%    -0.60%  (p=0.002 n=10+10)
GenericBuffer/paddedBooleanColumn/go1.18       590ns ± 0%   222ns ± 0%   -62.34%  (p=0.000 n=10+10)

name                                          old row/s    new row/s    delta
GenericBuffer/benchmarkRowType/go1.18           669M ± 0%   1340M ± 0%  +100.25%  (p=0.000 n=10+9)
GenericBuffer/booleanColumn/go1.18             1.70G ± 0%   3.17G ± 0%   +86.53%  (p=0.000 n=9+9)
GenericBuffer/int32Column/go1.18               1.74G ± 0%   4.25G ± 0%  +144.85%  (p=0.000 n=9+10)
GenericBuffer/int64Column/go1.18               1.80G ± 0%   3.28G ± 0%   +81.69%  (p=0.000 n=9+9)
GenericBuffer/floatColumn/go1.18               1.62G ± 0%   4.25G ± 0%  +161.58%  (p=0.000 n=10+10)
GenericBuffer/doubleColumn/go1.18              1.62G ± 0%   3.28G ± 0%  +101.82%  (p=0.000 n=9+8)
GenericBuffer/byteArrayColumn/go1.18            101M ± 1%    101M ± 2%      ~     (p=0.796 n=10+10)
GenericBuffer/fixedLenByteArrayColumn/go1.18    184M ± 0%    182M ± 0%    -0.99%  (p=0.000 n=10+9)
GenericBuffer/stringColumn/go1.18               161M ± 0%    160M ± 0%    -0.74%  (p=0.000 n=9+8)
GenericBuffer/indexedStringColumn/go1.18       24.9M ± 0%   24.9M ± 1%      ~     (p=0.370 n=8+9)
GenericBuffer/uuidColumn/go1.18                1.13G ± 0%   2.56G ± 0%  +127.41%  (p=0.000 n=10+10)
GenericBuffer/mapColumn/go1.18                 1.44M ± 0%   1.42M ± 1%    -1.68%  (p=0.000 n=9+10)
GenericBuffer/decimalColumn/go1.18             1.73G ± 0%   3.27G ± 0%   +88.23%  (p=0.000 n=10+10)
GenericBuffer/contact/go1.18                   76.8M ± 1%   77.3M ± 0%    +0.60%  (p=0.002 n=10+10)
GenericBuffer/paddedBooleanColumn/go1.18       1.70G ± 0%   4.50G ± 0%  +165.56%  (p=0.000 n=10+10)
```

I also refactored the tests that were using `quick.Check`, the function did not ignore unexported struct fields (even if they were named `_`) which caused panics when it attempted to set the values through reflection. I moved the code from `internal/bits` to `internal/quick` and made it more configurable to use throughout the project.